split up ast.hpp
