+++ src/query/frontend/ast/ast.hpp
@@ -17,7 +17,16 @@
 #include <vector>
 
 #include "query/exceptions.hpp"
+#include "query/frontend/ast/ast_storage.hpp"
 #include "query/frontend/ast/ast_visitor.hpp"
+#include "query/frontend/ast/ordering.hpp"
+#include "query/frontend/ast/query/binary_operator.hpp"
+#include "query/frontend/ast/query/expression.hpp"
+#include "query/frontend/ast/query/identifier.hpp"
+#include "query/frontend/ast/query/named_expression.hpp"
+#include "query/frontend/ast/query/pattern.hpp"
+#include "query/frontend/ast/query/query.hpp"
+#include "query/frontend/ast/query/where.hpp"
 #include "query/frontend/semantic/symbol.hpp"
 #include "query/interpret/awesome_memgraph_functions.hpp"
 #include "query/procedure/module_fwd.hpp"
@@ -44,193 +53,6 @@ inline constexpr std::string_view kCapacity = "capacity";
 inline constexpr std::string_view kResizeCoefficient = "resize_coefficient";
 inline constexpr std::uint16_t kDefaultResizeCoefficient = 2;
 inline constexpr std::string_view kDefaultMetric = "l2sq";
-
-struct LabelIx {
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const { return kType; }
-  friend bool operator==(const LabelIx &a, const LabelIx &b) { return a.ix == b.ix; }
-  friend bool operator<(const LabelIx &a, const LabelIx &b) { return a.ix < b.ix; }
-
-  std::string name;
-  int64_t ix;
-};
-
-struct PropertyIx {
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const { return kType; }
-  friend bool operator==(const PropertyIx &a, const PropertyIx &b) { return a.ix == b.ix; }
-  friend bool operator<(const PropertyIx &a, const PropertyIx &b) { return a.ix < b.ix; }
-
-  std::string name;
-  int64_t ix;
-};
-
-struct EdgeTypeIx {
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const { return kType; }
-  friend bool operator==(const EdgeTypeIx &a, const EdgeTypeIx &b) { return a.ix == b.ix; }
-  friend bool operator<(const EdgeTypeIx &a, const EdgeTypeIx &b) { return a.ix < b.ix; }
-
-  std::string name;
-  int64_t ix;
-};
-
-}  // namespace memgraph::query
-
-namespace std {
-
-template <>
-struct hash<memgraph::query::LabelIx> {
-  size_t operator()(const memgraph::query::LabelIx &label) const { return label.ix; }
-};
-
-template <>
-struct hash<memgraph::query::PropertyIx> {
-  size_t operator()(const memgraph::query::PropertyIx &prop) const { return prop.ix; }
-};
-
-template <>
-struct hash<memgraph::query::EdgeTypeIx> {
-  size_t operator()(const memgraph::query::EdgeTypeIx &edge_type) const { return edge_type.ix; }
-};
-
-}  // namespace std
-
-namespace memgraph::query {
-
-class Tree;
-
-// It would be better to call this AstTree, but we already have a class Tree,
-// which could be renamed to Node or AstTreeNode, but we also have a class
-// called NodeAtom...
-class AstStorage {
- public:
-  AstStorage() = default;
-  AstStorage(const AstStorage &) = delete;
-  AstStorage &operator=(const AstStorage &) = delete;
-  AstStorage(AstStorage &&) = default;
-  AstStorage &operator=(AstStorage &&) = default;
-
-  template <typename T, typename... Args>
-  T *Create(Args &&...args) {
-    T *ptr = new T(std::forward<Args>(args)...);
-    std::unique_ptr<T> tmp(ptr);
-    storage_.emplace_back(std::move(tmp));
-    return ptr;
-  }
-
-  LabelIx GetLabelIx(const std::string &name) { return LabelIx{name, FindOrAddName(name, &labels_)}; }
-
-  PropertyIx GetPropertyIx(const std::string &name) { return PropertyIx{name, FindOrAddName(name, &properties_)}; }
-
-  EdgeTypeIx GetEdgeTypeIx(const std::string &name) { return EdgeTypeIx{name, FindOrAddName(name, &edge_types_)}; }
-
-  // TODO: would be good if these were stable memory locations, then *Ix could have string_view rather than stringq
-  std::vector<std::string> labels_;
-  std::vector<std::string> edge_types_;
-  std::vector<std::string> properties_;
-
-  // Public only for serialization access
-  std::vector<std::unique_ptr<Tree>> storage_;
-
- private:
-  int64_t FindOrAddName(const std::string &name, std::vector<std::string> *names) {
-    for (int64_t i = 0; i < names->size(); ++i) {
-      if ((*names)[i] == name) {
-        return i;
-      }
-    }
-    names->push_back(name);
-    return names->size() - 1;
-  }
-};
-
-class Tree {
- public:
-  static const utils::TypeInfo kType;
-  virtual const utils::TypeInfo &GetTypeInfo() const { return kType; }
-
-  Tree() = default;
-  virtual ~Tree() = default;
-
-  virtual Tree *Clone(AstStorage *storage) const = 0;
-
- private:
-  friend class AstStorage;
-};
-
-class Expression : public memgraph::query::Tree,
-                   public utils::Visitable<HierarchicalTreeVisitor>,
-                   public utils::Visitable<ExpressionVisitor<TypedValue>>,
-                   public utils::Visitable<ExpressionVisitor<TypedValue *>>,
-                   public utils::Visitable<ExpressionVisitor<void>> {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  using utils::Visitable<HierarchicalTreeVisitor>::Accept;
-  using utils::Visitable<ExpressionVisitor<TypedValue>>::Accept;
-  using utils::Visitable<ExpressionVisitor<TypedValue *>>::Accept;
-  using utils::Visitable<ExpressionVisitor<void>>::Accept;
-
-  Expression() = default;
-
-  Expression *Clone(AstStorage *storage) const override = 0;
-
- private:
-  friend class AstStorage;
-};
-
-class Where : public memgraph::query::Tree, public utils::Visitable<HierarchicalTreeVisitor> {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  using utils::Visitable<HierarchicalTreeVisitor>::Accept;
-
-  Where() = default;
-
-  bool Accept(HierarchicalTreeVisitor &visitor) override {
-    if (visitor.PreVisit(*this)) {
-      expression_->Accept(visitor);
-    }
-    return visitor.PostVisit(*this);
-  }
-
-  memgraph::query::Expression *expression_{nullptr};
-
-  Where *Clone(AstStorage *storage) const override {
-    Where *object = storage->Create<Where>();
-    object->expression_ = expression_ ? expression_->Clone(storage) : nullptr;
-    return object;
-  }
-
- protected:
-  explicit Where(Expression *expression) : expression_(expression) {}
-
- private:
-  friend class AstStorage;
-};
-
-class BinaryOperator : public memgraph::query::Expression {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  BinaryOperator() = default;
-
-  memgraph::query::Expression *expression1_{nullptr};
-  memgraph::query::Expression *expression2_{nullptr};
-
-  BinaryOperator *Clone(AstStorage *storage) const override = 0;
-
- protected:
-  BinaryOperator(Expression *expression1, Expression *expression2)
-      : expression1_(expression1), expression2_(expression2) {}
-
- private:
-  friend class AstStorage;
-};
 
 class UnaryOperator : public memgraph::query::Expression {
  public:
@@ -884,78 +706,6 @@ class IsNullOperator : public memgraph::query::UnaryOperator {
   friend class AstStorage;
 };
 
-class Aggregation : public memgraph::query::BinaryOperator {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  enum class Op { COUNT, MIN, MAX, SUM, AVG, COLLECT_LIST, COLLECT_MAP, PROJECT_PATH, PROJECT_LISTS };
-
-  Aggregation() = default;
-
-  static const constexpr char *const kCount = "COUNT";
-  static const constexpr char *const kMin = "MIN";
-  static const constexpr char *const kMax = "MAX";
-  static const constexpr char *const kSum = "SUM";
-  static const constexpr char *const kAvg = "AVG";
-  static const constexpr char *const kCollect = "COLLECT";
-  static const constexpr char *const kProject = "PROJECT";
-
-  static std::string OpToString(Op op) {
-    const char *op_strings[] = {kCount, kMin, kMax, kSum, kAvg, kCollect, kCollect, kProject, kProject};
-    return op_strings[static_cast<int>(op)];
-  }
-
-  DEFVISITABLE(ExpressionVisitor<TypedValue>);
-  DEFVISITABLE(ExpressionVisitor<TypedValue *>);
-  DEFVISITABLE(ExpressionVisitor<void>);
-  bool Accept(HierarchicalTreeVisitor &visitor) override {
-    if (visitor.PreVisit(*this)) {
-      if (expression1_) expression1_->Accept(visitor);
-      if (expression2_) expression2_->Accept(visitor);
-    }
-    return visitor.PostVisit(*this);
-  }
-
-  Aggregation *MapTo(const Symbol &symbol) {
-    symbol_pos_ = symbol.position();
-    return this;
-  }
-
-  memgraph::query::Aggregation::Op op_;
-  /// Symbol table position of the symbol this Aggregation is mapped to.
-  int32_t symbol_pos_{-1};
-  bool distinct_{false};
-
-  Aggregation *Clone(AstStorage *storage) const override {
-    Aggregation *object = storage->Create<Aggregation>();
-    object->expression1_ = expression1_ ? expression1_->Clone(storage) : nullptr;
-    object->expression2_ = expression2_ ? expression2_->Clone(storage) : nullptr;
-    object->op_ = op_;
-    object->symbol_pos_ = symbol_pos_;
-    object->distinct_ = distinct_;
-    return object;
-  }
-
- protected:
-  // Use only for serialization.
-  explicit Aggregation(Op op) : op_(op) {}
-
-  /// Aggregation's first expression is the value being aggregated. The second expression is used either as a key in
-  /// COLLECT_MAP or for the relationships list in the two-argument overload of PROJECT_PATH; no other aggregate
-  /// functions use this parameter.
-  Aggregation(Expression *expression1, Expression *expression2, Op op, bool distinct)
-      : BinaryOperator(expression1, expression2), op_(op), distinct_(distinct) {
-    // COUNT without expression denotes COUNT(*) in cypher.
-    DMG_ASSERT(expression1 || op == Aggregation::Op::COUNT, "All aggregations, except COUNT require expression1");
-    DMG_ASSERT((expression2 == nullptr) ^ (op == Aggregation::Op::PROJECT_LISTS || op == Aggregation::Op::COLLECT_MAP),
-               "expression2 is obligatory in COLLECT_MAP and PROJECT_LISTS, and invalid otherwise");
-  }
-
- private:
-  friend class AstStorage;
-};
-
 class ListSlicingOperator : public memgraph::query::Expression {
  public:
   static const utils::TypeInfo kType;
@@ -1209,43 +959,6 @@ class MapProjectionLiteral : public memgraph::query::BaseLiteral {
  protected:
   explicit MapProjectionLiteral(Expression *map_variable, std::unordered_map<PropertyIx, Expression *> &&elements)
       : map_variable_(map_variable), elements_(std::move(elements)) {}
-
- private:
-  friend class AstStorage;
-};
-
-class Identifier : public memgraph::query::Expression {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  Identifier() = default;
-
-  DEFVISITABLE(ExpressionVisitor<TypedValue>);
-  DEFVISITABLE(ExpressionVisitor<TypedValue *>);
-  DEFVISITABLE(ExpressionVisitor<void>);
-  DEFVISITABLE(HierarchicalTreeVisitor);
-
-  Identifier *MapTo(const Symbol &symbol) {
-    symbol_pos_ = symbol.position();
-    return this;
-  }
-
-  explicit Identifier(const std::string &name) : name_(name) {}
-  Identifier(const std::string &name, bool user_declared) : name_(name), user_declared_(user_declared) {}
-
-  std::string name_;
-  bool user_declared_{true};
-  /// Symbol table position of the symbol this Identifier is mapped to.
-  int32_t symbol_pos_{-1};
-
-  Identifier *Clone(AstStorage *storage) const override {
-    Identifier *object = storage->Create<Identifier>();
-    object->name_ = name_;
-    object->user_declared_ = user_declared_;
-    object->symbol_pos_ = symbol_pos_;
-    return object;
-  }
 
  private:
   friend class AstStorage;
@@ -1839,86 +1552,6 @@ class RegexMatch : public memgraph::query::Expression {
   RegexMatch(Expression *string_expr, Expression *regex) : string_expr_(string_expr), regex_(regex) {}
 };
 
-class NamedExpression : public memgraph::query::Tree,
-                        public utils::Visitable<HierarchicalTreeVisitor>,
-                        public utils::Visitable<ExpressionVisitor<TypedValue>>,
-                        public utils::Visitable<ExpressionVisitor<TypedValue *>>,
-                        public utils::Visitable<ExpressionVisitor<void>> {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  using utils::Visitable<ExpressionVisitor<TypedValue>>::Accept;
-  using utils::Visitable<ExpressionVisitor<void>>::Accept;
-  using utils::Visitable<HierarchicalTreeVisitor>::Accept;
-
-  NamedExpression() = default;
-
-  DEFVISITABLE(ExpressionVisitor<TypedValue>);
-  DEFVISITABLE(ExpressionVisitor<TypedValue *>);
-  DEFVISITABLE(ExpressionVisitor<void>);
-  bool Accept(HierarchicalTreeVisitor &visitor) override {
-    if (visitor.PreVisit(*this)) {
-      expression_->Accept(visitor);
-    }
-    return visitor.PostVisit(*this);
-  }
-
-  NamedExpression *MapTo(const Symbol &symbol) {
-    symbol_pos_ = symbol.position();
-    return this;
-  }
-
-  std::string name_;
-  memgraph::query::Expression *expression_{nullptr};
-  /// This field contains token position of first token in named expression used to create name_. If NamedExpression
-  /// object is not created from query or it is aliased leave this value at -1.
-  int32_t token_position_{-1};
-  /// Symbol table position of the symbol this NamedExpression is mapped to.
-  int32_t symbol_pos_{-1};
-  /// True if the variable is aliased
-  bool is_aliased_{false};
-
-  NamedExpression *Clone(AstStorage *storage) const override {
-    NamedExpression *object = storage->Create<NamedExpression>();
-    object->name_ = name_;
-    object->expression_ = expression_ ? expression_->Clone(storage) : nullptr;
-    object->token_position_ = token_position_;
-    object->symbol_pos_ = symbol_pos_;
-    object->is_aliased_ = is_aliased_;
-    return object;
-  }
-
- protected:
-  explicit NamedExpression(const std::string &name) : name_(name) {}
-  NamedExpression(const std::string &name, Expression *expression) : name_(name), expression_(expression) {}
-  NamedExpression(const std::string &name, Expression *expression, int token_position)
-      : name_(name), expression_(expression), token_position_(token_position) {}
-
- private:
-  friend class AstStorage;
-};
-
-class PatternAtom : public memgraph::query::Tree, public utils::Visitable<HierarchicalTreeVisitor> {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  using utils::Visitable<HierarchicalTreeVisitor>::Accept;
-
-  PatternAtom() = default;
-
-  memgraph::query::Identifier *identifier_{nullptr};
-
-  PatternAtom *Clone(AstStorage *storage) const override = 0;
-
- protected:
-  explicit PatternAtom(Identifier *identifier) : identifier_(identifier) {}
-
- private:
-  friend class AstStorage;
-};
-
 class NodeAtom : public memgraph::query::PatternAtom {
  public:
   static const utils::TypeInfo kType;
@@ -2114,44 +1747,6 @@ class EdgeAtom : public memgraph::query::PatternAtom {
   friend class AstStorage;
 };
 
-class Pattern : public memgraph::query::Tree, public utils::Visitable<HierarchicalTreeVisitor> {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  using utils::Visitable<HierarchicalTreeVisitor>::Accept;
-
-  Pattern() = default;
-
-  bool Accept(HierarchicalTreeVisitor &visitor) override {
-    if (visitor.PreVisit(*this)) {
-      bool cont = identifier_->Accept(visitor);
-      for (auto &part : atoms_) {
-        if (cont) {
-          cont = part->Accept(visitor);
-        }
-      }
-    }
-    return visitor.PostVisit(*this);
-  }
-
-  memgraph::query::Identifier *identifier_{nullptr};
-  std::vector<memgraph::query::PatternAtom *> atoms_;
-
-  Pattern *Clone(AstStorage *storage) const override {
-    Pattern *object = storage->Create<Pattern>();
-    object->identifier_ = identifier_ ? identifier_->Clone(storage) : nullptr;
-    object->atoms_.resize(atoms_.size());
-    for (auto i3 = 0; i3 < atoms_.size(); ++i3) {
-      object->atoms_[i3] = atoms_[i3] ? atoms_[i3]->Clone(storage) : nullptr;
-    }
-    return object;
-  }
-
- private:
-  friend class AstStorage;
-};
-
 class Clause : public memgraph::query::Tree, public utils::Visitable<HierarchicalTreeVisitor> {
  public:
   static const utils::TypeInfo kType;
@@ -2236,21 +1831,6 @@ class CypherUnion : public memgraph::query::Tree, public utils::Visitable<Hierar
   explicit CypherUnion(bool distinct) : distinct_(distinct) {}
   CypherUnion(bool distinct, SingleQuery *single_query, std::vector<Symbol> union_symbols)
       : single_query_(single_query), distinct_(distinct), union_symbols_(union_symbols) {}
-
- private:
-  friend class AstStorage;
-};
-
-class Query : public memgraph::query::Tree, public utils::Visitable<QueryVisitor<void>> {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  using utils::Visitable<QueryVisitor<void>>::Accept;
-
-  Query() = default;
-
-  Query *Clone(AstStorage *storage) const override = 0;
 
  private:
   friend class AstStorage;
@@ -2699,9 +2279,6 @@ class Match : public memgraph::query::Clause {
   friend class AstStorage;
 };
 
-/// Defines the order for sorting values (ascending or descending).
-enum class Ordering { ASC, DESC };
-
 struct SortItem {
   static const utils::TypeInfo kType;
   const utils::TypeInfo &GetTypeInfo() const { return kType; }
@@ -3140,161 +2717,6 @@ class Unwind : public memgraph::query::Clause {
 
  private:
   friend class AstStorage;
-};
-
-class AuthQuery : public memgraph::query::Query {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  enum class Action {
-    CREATE_ROLE,
-    DROP_ROLE,
-    SHOW_ROLES,
-    CREATE_USER,
-    SET_PASSWORD,
-    CHANGE_PASSWORD,
-    DROP_USER,
-    SHOW_CURRENT_USER,
-    SHOW_USERS,
-    SET_ROLE,
-    CLEAR_ROLE,
-    GRANT_PRIVILEGE,
-    DENY_PRIVILEGE,
-    REVOKE_PRIVILEGE,
-    SHOW_PRIVILEGES,
-    SHOW_ROLE_FOR_USER,
-    SHOW_USERS_FOR_ROLE,
-    GRANT_DATABASE_TO_USER,
-    DENY_DATABASE_FROM_USER,
-    REVOKE_DATABASE_FROM_USER,
-    SHOW_DATABASE_PRIVILEGES,
-    SET_MAIN_DATABASE,
-    GRANT_IMPERSONATE_USER,
-    DENY_IMPERSONATE_USER,
-  };
-
-  enum class Privilege {
-    CREATE,
-    DELETE,
-    MATCH,
-    MERGE,
-    SET,
-    REMOVE,
-    INDEX,
-    STATS,
-    AUTH,
-    CONSTRAINT,
-    DUMP,
-    REPLICATION,
-    DURABILITY,
-    READ_FILE,
-    FREE_MEMORY,
-    TRIGGER,
-    CONFIG,
-    STREAM,
-    MODULE_READ,
-    MODULE_WRITE,
-    WEBSOCKET,
-    STORAGE_MODE,
-    TRANSACTION_MANAGEMENT,
-    MULTI_DATABASE_EDIT,
-    MULTI_DATABASE_USE,
-    COORDINATOR,
-    IMPERSONATE_USER,
-  };
-
-  enum class FineGrainedPrivilege { NOTHING, READ, UPDATE, CREATE_DELETE };
-
-  AuthQuery() = default;
-
-  DEFVISITABLE(QueryVisitor<void>);
-
-  memgraph::query::AuthQuery::Action action_;
-  std::string user_;
-  std::string role_;
-  std::string user_or_role_;
-  memgraph::query::Expression *old_password_{nullptr};
-  memgraph::query::Expression *new_password_{nullptr};
-  bool if_not_exists_;
-  memgraph::query::Expression *password_{nullptr};
-  std::string database_;
-  std::vector<memgraph::query::AuthQuery::Privilege> privileges_;
-  std::vector<std::unordered_map<memgraph::query::AuthQuery::FineGrainedPrivilege, std::vector<std::string>>>
-      label_privileges_;
-  std::vector<std::unordered_map<memgraph::query::AuthQuery::FineGrainedPrivilege, std::vector<std::string>>>
-      edge_type_privileges_;
-  std::vector<std::string> impersonation_targets_;
-
-  AuthQuery *Clone(AstStorage *storage) const override {
-    auto *object = storage->Create<AuthQuery>();
-    object->action_ = action_;
-    object->user_ = user_;
-    object->role_ = role_;
-    object->user_or_role_ = user_or_role_;
-    object->old_password_ = old_password_;
-    object->new_password_ = new_password_;
-    object->if_not_exists_ = if_not_exists_;
-    object->password_ = password_ ? password_->Clone(storage) : nullptr;
-    object->database_ = database_;
-    object->privileges_ = privileges_;
-    object->label_privileges_ = label_privileges_;
-    object->edge_type_privileges_ = edge_type_privileges_;
-    object->impersonation_targets_ = impersonation_targets_;
-    return object;
-  }
-
- protected:
-  AuthQuery(Action action, std::string user, std::string role, std::string user_or_role, bool if_not_exists,
-            Expression *password, std::string database, std::vector<Privilege> privileges,
-            std::vector<std::unordered_map<FineGrainedPrivilege, std::vector<std::string>>> label_privileges,
-            std::vector<std::unordered_map<FineGrainedPrivilege, std::vector<std::string>>> edge_type_privileges,
-            std::vector<std::string> impersonation_targets)
-      : action_(action),
-        user_(user),
-        role_(role),
-        user_or_role_(user_or_role),
-        if_not_exists_(if_not_exists),
-        password_(password),
-        database_(database),
-        privileges_(privileges),
-        label_privileges_(label_privileges),
-        edge_type_privileges_(edge_type_privileges),
-        impersonation_targets_(impersonation_targets) {}
-
- private:
-  friend class AstStorage;
-};
-
-/// Constant that holds all available privileges.
-const std::vector<AuthQuery::Privilege> kPrivilegesAll = {
-    AuthQuery::Privilege::CREATE,
-    AuthQuery::Privilege::DELETE,
-    AuthQuery::Privilege::MATCH,
-    AuthQuery::Privilege::MERGE,
-    AuthQuery::Privilege::SET,
-    AuthQuery::Privilege::REMOVE,
-    AuthQuery::Privilege::INDEX,
-    AuthQuery::Privilege::STATS,
-    AuthQuery::Privilege::AUTH,
-    AuthQuery::Privilege::CONSTRAINT,
-    AuthQuery::Privilege::DUMP,
-    AuthQuery::Privilege::REPLICATION,
-    AuthQuery::Privilege::READ_FILE,
-    AuthQuery::Privilege::DURABILITY,
-    AuthQuery::Privilege::FREE_MEMORY,
-    AuthQuery::Privilege::TRIGGER,
-    AuthQuery::Privilege::CONFIG,
-    AuthQuery::Privilege::STREAM,
-    AuthQuery::Privilege::MODULE_READ,
-    AuthQuery::Privilege::MODULE_WRITE,
-    AuthQuery::Privilege::WEBSOCKET,
-    AuthQuery::Privilege::TRANSACTION_MANAGEMENT,
-    AuthQuery::Privilege::STORAGE_MODE,
-    AuthQuery::Privilege::MULTI_DATABASE_EDIT,
-    AuthQuery::Privilege::MULTI_DATABASE_USE,
-    AuthQuery::Privilege::COORDINATOR,
-    AuthQuery::Privilege::IMPERSONATE_USER,
 };
 
 class DatabaseInfoQuery : public memgraph::query::Query {
@@ -3971,106 +3393,6 @@ class AnalyzeGraphQuery : public memgraph::query::Query {
     object->labels_ = labels_;
     return object;
   }
-};
-
-class Exists : public memgraph::query::Expression {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  Exists() = default;
-
-  DEFVISITABLE(ExpressionVisitor<TypedValue>);
-  DEFVISITABLE(ExpressionVisitor<TypedValue *>);
-  DEFVISITABLE(ExpressionVisitor<void>);
-  bool Accept(HierarchicalTreeVisitor &visitor) override {
-    if (visitor.PreVisit(*this)) {
-      pattern_->Accept(visitor);
-    }
-    return visitor.PostVisit(*this);
-  }
-  Exists *MapTo(const Symbol &symbol) {
-    symbol_pos_ = symbol.position();
-    return this;
-  }
-
-  memgraph::query::Pattern *pattern_{nullptr};
-  /// Symbol table position of the symbol this Aggregation is mapped to.
-  int32_t symbol_pos_{-1};
-
-  Exists *Clone(AstStorage *storage) const override {
-    Exists *object = storage->Create<Exists>();
-    object->pattern_ = pattern_ ? pattern_->Clone(storage) : nullptr;
-    object->symbol_pos_ = symbol_pos_;
-    return object;
-  }
-
- protected:
-  Exists(Pattern *pattern) : pattern_(pattern) {}
-
- private:
-  friend class AstStorage;
-};
-
-class PatternComprehension : public memgraph::query::Expression {
- public:
-  static const utils::TypeInfo kType;
-  const utils::TypeInfo &GetTypeInfo() const override { return kType; }
-
-  PatternComprehension() = default;
-
-  DEFVISITABLE(ExpressionVisitor<TypedValue>);
-  DEFVISITABLE(ExpressionVisitor<TypedValue *>);
-  DEFVISITABLE(ExpressionVisitor<void>);
-
-  bool Accept(HierarchicalTreeVisitor &visitor) override {
-    if (visitor.PreVisit(*this)) {
-      if (variable_) {
-        variable_->Accept(visitor);
-      }
-      pattern_->Accept(visitor);
-      if (filter_) {
-        filter_->Accept(visitor);
-      }
-      resultExpr_->Accept(visitor);
-    }
-    return visitor.PostVisit(*this);
-  }
-
-  PatternComprehension *MapTo(const Symbol &symbol) {
-    symbol_pos_ = symbol.position();
-    return this;
-  }
-
-  // The variable name.
-  Identifier *variable_{nullptr};
-  // The pattern to match.
-  Pattern *pattern_{nullptr};
-  // Optional WHERE clause for filtering.
-  Where *filter_{nullptr};
-  // The projection expression.
-  Expression *resultExpr_{nullptr};
-
-  /// Symbol table position of the symbol this Aggregation is mapped to.
-  int32_t symbol_pos_{-1};
-
-  PatternComprehension *Clone(AstStorage *storage) const override {
-    auto *object = storage->Create<PatternComprehension>();
-    object->variable_ = variable_ ? variable_->Clone(storage) : nullptr;
-    object->pattern_ = pattern_ ? pattern_->Clone(storage) : nullptr;
-    object->filter_ = filter_ ? filter_->Clone(storage) : nullptr;
-    object->resultExpr_ = resultExpr_ ? resultExpr_->Clone(storage) : nullptr;
-
-    object->symbol_pos_ = symbol_pos_;
-    return object;
-  }
-
- protected:
-  PatternComprehension(Identifier *variable, Pattern *pattern, Where *filter, Expression *resultExpr)
-      : variable_(variable), pattern_(pattern), filter_(filter), resultExpr_(resultExpr) {}
-
- private:
-  friend class AstStorage;
 };
 
 class CallSubquery : public memgraph::query::Clause {
