+++ src/query/frontend/ast/ast_visitor.hpp
@@ -10,8 +10,6 @@
 // licenses/APL.txt.
 
 #pragma once
-// NOLINTNEXTLINE(misc-header-include-cycle)
-#include "query/frontend/ast/ast.hpp"
 #include "utils/visitor.hpp"
 
 namespace memgraph::query {
