+++ src/query/frontend/ast/cypher_main_visitor.hpp
@@ -17,7 +17,6 @@
 #include "query/frontend/opencypher/generated/MemgraphCypherBaseVisitor.h"
 #pragma pop_macro("EOF")  // bring EOF back
 
-#include "query/frontend/ast/ast.hpp"
 #include "query/parameters.hpp"
 #include "utils/exceptions.hpp"
 #include "utils/logging.hpp"
