+++ src/query/plan/cost_estimator.hpp
@@ -11,7 +11,6 @@
 
 #pragma once
 
-#include "query/frontend/ast/ast.hpp"
 #include "query/parameters.hpp"
 #include "query/plan/operator.hpp"
 #include "query/plan/rewrite/index_lookup.hpp"
