+++ src/query/plan/rewrite/general.hpp
@@ -1,4 +1,4 @@
-// Copyright 2024 Memgraph Ltd.
+// Copyright 2025 Memgraph Ltd.
 //
 // Use of this software is governed by the Business Source License
 // included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
@@ -18,7 +18,7 @@
 
 #include <unordered_set>
 
-#include "query/frontend/ast/ast.hpp"
+#include "query/frontend/ast/query/expression.hpp"
 
 namespace memgraph::query::plan {
