+++ src/query/plan/rewrite/range.hpp
@@ -1,4 +1,4 @@
-// Copyright 2024 Memgraph Ltd.
+// Copyright 2025 Memgraph Ltd.
 //
 // Use of this software is governed by the Business Source License
 // included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
@@ -11,7 +11,6 @@
 
 #pragma once
 
-#include "query/frontend/ast/ast.hpp"
 #include "query/plan/preprocess.hpp"
 
 namespace memgraph::query::plan {
