+++ src/query/plan/rule_based_planner.hpp
@@ -18,7 +18,6 @@
 
 #include "flags/run_time_configurable.hpp"
 #include "query/database_access.hpp"
-#include "query/frontend/ast/ast.hpp"
 #include "query/frontend/ast/ast_visitor.hpp"
 #include "query/plan/operator.hpp"
 #include "query/plan/preprocess.hpp"
