+++ src/query/trigger.hpp
@@ -1,4 +1,4 @@
-// Copyright 2024 Memgraph Ltd.
+// Copyright 2025 Memgraph Ltd.
 //
 // Use of this software is governed by the Business Source License
 // included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
@@ -23,7 +23,6 @@
 #include "query/config.hpp"
 #include "query/cypher_query_interpreter.hpp"
 #include "query/database_access.hpp"
-#include "query/frontend/ast/ast.hpp"
 #include "query/trigger_context.hpp"
 #include "storage/v2/property_value.hpp"
 #include "utils/rw_spin_lock.hpp"
