+++ src/utils/frame_change_id.hpp
@@ -1,4 +1,4 @@
-// Copyright 2023 Memgraph Ltd.
+// Copyright 2025 Memgraph Ltd.
 //
 // Use of this software is governed by the Business Source License
 // included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
@@ -11,7 +11,6 @@
 
 #include <string>
 
-#include "query/frontend/ast/ast.hpp"
 #include "spdlog/spdlog.h"
 
 namespace memgraph::utils {
